move model into its own sub project fix the license text
