--- core/src/main/scala/com/qvantel/jsonapi/model/package.scala
@@ -1,43 +1,0 @@
-/*
-Copyright (c) 2017, Qvantel
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without
-modification, are permitted provided that the following conditions are met:
- * Redistributions of source code must retain the above copyright
-      notice, this list of conditions and the following disclaimer.
- * Redistributions in binary form must reproduce the above copyright
-      notice, this list of conditions and the following disclaimer in the
-      documentation and/or other materials provided with the distribution.
- * Neither the name of the <organization> nor the
-      names of its contributors may be used to endorse or promote products
-      derived from this software without specific prior written permission.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-DISCLAIMED. IN NO EVENT SHALL <COPYRIGHT HOLDER> BE LIABLE FOR ANY
-DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-(INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package com.qvantel.jsonapi
-
-import _root_.spray.json._
-
-package object model {
-  type MetaObject    = Map[String, JsValue]
-  type Links         = Map[String, Link]
-  type Attributes    = Map[String, JsValue]
-  type Relationships = Map[String, RelationshipObject]
-
-  final implicit class JsonMapOps(val underlying: Map[String, JsValue]) extends AnyVal {
-    @inline def getAs[A](name: String)(implicit r: JsonReader[Option[A]]): Option[A] =
-      underlying.get(name).flatMap(_.convertTo[Option[A]])
-
-    @inline def getAs[A](name: Symbol)(implicit r: JsonReader[Option[A]]): Option[A] = getAs(name.name)
-  }
-}
