+++ model/src/main/scala/com/qvantel/jsonapi/model/JsonApiInfo.scala
@@ -9,14 +9,14 @@ modification, are permitted provided that the following conditions are met:
  * Redistributions in binary form must reproduce the above copyright
       notice, this list of conditions and the following disclaimer in the
       documentation and/or other materials provided with the distribution.
- * Neither the name of the <organization> nor the
+ * Neither the name of the Qvantel nor the
       names of its contributors may be used to endorse or promote products
       derived from this software without specific prior written permission.
 
 THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
 ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
 WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-DISCLAIMED. IN NO EVENT SHALL <COPYRIGHT HOLDER> BE LIABLE FOR ANY
+DISCLAIMED. IN NO EVENT SHALL Qvantel BE LIABLE FOR ANY
 DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
 (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
 LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
@@ -29,29 +29,20 @@ package com.qvantel.jsonapi.model
 import _root_.spray.json.DefaultJsonProtocol._
 import _root_.spray.json._
 
-final case class ResourceIdentifierObject(id: String, `type`: String, meta: MetaObject)
+final case class JsonApiInfo(version: Option[String], meta: MetaObject)
 
-object ResourceIdentifierObject {
-  implicit object ResourceIdentifierObjectJsonFormat extends RootJsonFormat[ResourceIdentifierObject] {
-    override def write(obj: ResourceIdentifierObject): JsValue = obj match {
-      case ResourceIdentifierObject(id, t, meta) if meta.isEmpty => JsObject("id" -> id.toJson, "type" -> t.toJson)
-      case ResourceIdentifierObject(id, t, meta) =>
-        JsObject("id" -> id.toJson, "type" -> t.toJson, "meta" -> meta.toJson)
+object JsonApiInfo {
+  implicit object JsonApiInfoJsonFormat extends RootJsonFormat[JsonApiInfo] {
+    override def write(obj: JsonApiInfo): JsValue = obj match {
+      case JsonApiInfo(None, meta) if meta.isEmpty          => JsObject.empty
+      case JsonApiInfo(Some(version), meta) if meta.isEmpty => JsObject("version" -> version.toJson)
+      case JsonApiInfo(Some(version), meta)                 => JsObject("version" -> version.toJson, "meta" -> meta.toJson)
     }
 
-    override def read(json: JsValue): ResourceIdentifierObject = {
+    override def read(json: JsValue): JsonApiInfo = {
       val fields = json.asJsObject.fields
-      ResourceIdentifierObject(
-        id = fields
-          .get("id")
-          .map(_.convertTo[String])
-          .getOrElse(deserializationError(s"Expected an ‘id’ field in resource identifier object")),
-        `type` = fields
-          .get("type")
-          .map(_.convertTo[String])
-          .getOrElse(deserializationError(s"Expected an ‘type’ field in resource identifier object")),
-        meta = fields.get("meta").map(_.convertTo[MetaObject]).getOrElse(Map.empty)
-      )
+      JsonApiInfo(version = fields.get("version").flatMap(_.convertTo[Option[String]]),
+                  meta = fields.get("meta").map(_.convertTo[MetaObject]).getOrElse(Map.empty))
     }
   }
 }
