+++ model/src/main/scala/com/qvantel/jsonapi/model/Link.scala
@@ -9,14 +9,14 @@ modification, are permitted provided that the following conditions are met:
  * Redistributions in binary form must reproduce the above copyright
       notice, this list of conditions and the following disclaimer in the
       documentation and/or other materials provided with the distribution.
- * Neither the name of the <organization> nor the
+ * Neither the name of the Qvantel nor the
       names of its contributors may be used to endorse or promote products
       derived from this software without specific prior written permission.
 
 THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
 ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
 WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-DISCLAIMED. IN NO EVENT SHALL <COPYRIGHT HOLDER> BE LIABLE FOR ANY
+DISCLAIMED. IN NO EVENT SHALL Qvantel BE LIABLE FOR ANY
 DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
 (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
 LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
@@ -28,25 +28,38 @@ package com.qvantel.jsonapi.model
 
 import _root_.spray.json.DefaultJsonProtocol._
 import _root_.spray.json._
+import _root_.spray.http.Uri
 
-sealed abstract class ResourceLinkage
+sealed abstract class Link {
+  def href: Uri
+}
 
-object ResourceLinkage {
-  final case class ToOne(resource: Option[ResourceIdentifierObject]) extends ResourceLinkage
-  final case class ToMany(resources: Set[ResourceIdentifierObject])  extends ResourceLinkage
+object Link {
+  final case class Url(href: Uri)                          extends Link
+  final case class LinkObject(href: Uri, meta: MetaObject) extends Link
 
-  implicit object ResourceLinkageJsonFormat extends JsonFormat[ResourceLinkage] {
-    override def write(obj: ResourceLinkage): JsValue = obj match {
-      case ToOne(None)           => JsNull
-      case ToOne(Some(resource)) => resource.toJson
-      case ToMany(resources)     => JsArray(resources.map(_.toJson).toVector)
+  implicit object LinkJsonFormat extends JsonFormat[Link] {
+    override def write(obj: Link): JsValue = obj match {
+      case Url(href) => JsString(href.toString)
+      case LinkObject(href, meta) =>
+        if (meta.nonEmpty) {
+          JsObject("href" -> href.toString.toJson, "meta" -> meta.toJson)
+        } else {
+          JsObject("href" -> href.toString.toJson)
+        }
     }
 
-    override def read(json: JsValue): ResourceLinkage = json match {
-      case JsNull      => ToOne(None)
-      case o: JsObject => ToOne(Some(json.convertTo[ResourceIdentifierObject]))
-      case a: JsArray  => ToMany(json.convertTo[Set[ResourceIdentifierObject]])
-      case invalid     => deserializationError(s"Invalid resource linkage: ‘$invalid’")
+    override def read(json: JsValue): Link = json match {
+      case JsString(href) => Url(href)
+      case JsObject(fields) =>
+        LinkObject(
+          href = fields.get("href").map(_.convertTo[String]).getOrElse(deserializationError("Expected ‘href’ in Link")),
+          meta = fields.get("meta").map(_.convertTo[MetaObject]).getOrElse(Map.empty)
+        )
+      case invalid => deserializationError(s"Expected Link as JsString or JsObject, got ‘$invalid’")
     }
   }
+
+  def convertToLinks: (JsValue) => Links =
+    _.convertTo[Map[String, JsValue]].filterNot(_._2 == JsNull).map(entry => (entry._1, entry._2.convertTo[Link]))
 }
