+++ model/src/main/scala/com/qvantel/jsonapi/model/ResourceLinkage.scala
@@ -9,14 +9,14 @@ modification, are permitted provided that the following conditions are met:
  * Redistributions in binary form must reproduce the above copyright
       notice, this list of conditions and the following disclaimer in the
       documentation and/or other materials provided with the distribution.
- * Neither the name of the <organization> nor the
+ * Neither the name of the Qvantel nor the
       names of its contributors may be used to endorse or promote products
       derived from this software without specific prior written permission.
 
 THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
 ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
 WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-DISCLAIMED. IN NO EVENT SHALL <COPYRIGHT HOLDER> BE LIABLE FOR ANY
+DISCLAIMED. IN NO EVENT SHALL Qvantel BE LIABLE FOR ANY
 DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
 (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
 LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
@@ -29,25 +29,24 @@ package com.qvantel.jsonapi.model
 import _root_.spray.json.DefaultJsonProtocol._
 import _root_.spray.json._
 
-final case class RelationshipObject(links: Links, data: Option[ResourceLinkage], meta: MetaObject)
+sealed abstract class ResourceLinkage
 
-object RelationshipObject {
-  implicit object RelationshipObjectJsonFormat extends RootJsonFormat[RelationshipObject] {
-    override def write(obj: RelationshipObject): JsValue = {
-      val builder = Map.newBuilder[String, JsValue]
-      if (obj.links.nonEmpty) builder += "links" -> obj.links.toJson
-      if (obj.data.isDefined) builder += "data"  -> obj.data.toJson
-      if (obj.meta.nonEmpty) builder += "meta"   -> obj.meta.toJson
-      JsObject(builder.result())
+object ResourceLinkage {
+  final case class ToOne(resource: Option[ResourceIdentifierObject]) extends ResourceLinkage
+  final case class ToMany(resources: Set[ResourceIdentifierObject])  extends ResourceLinkage
+
+  implicit object ResourceLinkageJsonFormat extends JsonFormat[ResourceLinkage] {
+    override def write(obj: ResourceLinkage): JsValue = obj match {
+      case ToOne(None)           => JsNull
+      case ToOne(Some(resource)) => resource.toJson
+      case ToMany(resources)     => JsArray(resources.map(_.toJson).toVector)
     }
 
-    override def read(json: JsValue): RelationshipObject = {
-      val fields = json.asJsObject.fields
-      RelationshipObject(
-        links = fields.get("links").map(Link.convertToLinks).getOrElse(Map.empty),
-        data = fields.get("data").map(_.convertTo[ResourceLinkage]),
-        meta = fields.get("meta").map(_.convertTo[MetaObject]).getOrElse(Map.empty)
-      )
+    override def read(json: JsValue): ResourceLinkage = json match {
+      case JsNull      => ToOne(None)
+      case o: JsObject => ToOne(Some(json.convertTo[ResourceIdentifierObject]))
+      case a: JsArray  => ToMany(json.convertTo[Set[ResourceIdentifierObject]])
+      case invalid     => deserializationError(s"Invalid resource linkage: ‘$invalid’")
     }
   }
 }
