+++ model/src/main/scala/com/qvantel/jsonapi/model/package.scala
@@ -9,14 +9,14 @@ modification, are permitted provided that the following conditions are met:
  * Redistributions in binary form must reproduce the above copyright
       notice, this list of conditions and the following disclaimer in the
       documentation and/or other materials provided with the distribution.
- * Neither the name of the <organization> nor the
+ * Neither the name of the Qvantel nor the
       names of its contributors may be used to endorse or promote products
       derived from this software without specific prior written permission.
 
 THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
 ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
 WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-DISCLAIMED. IN NO EVENT SHALL <COPYRIGHT HOLDER> BE LIABLE FOR ANY
+DISCLAIMED. IN NO EVENT SHALL Qvantel BE LIABLE FOR ANY
 DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
 (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
 LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
@@ -24,25 +24,20 @@ ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.qvantel.jsonapi.model
+package com.qvantel.jsonapi
 
-import _root_.spray.json.DefaultJsonProtocol._
 import _root_.spray.json._
 
-final case class JsonApiInfo(version: Option[String], meta: MetaObject)
+package object model {
+  type MetaObject    = Map[String, JsValue]
+  type Links         = Map[String, Link]
+  type Attributes    = Map[String, JsValue]
+  type Relationships = Map[String, RelationshipObject]
 
-object JsonApiInfo {
-  implicit object JsonApiInfoJsonFormat extends RootJsonFormat[JsonApiInfo] {
-    override def write(obj: JsonApiInfo): JsValue = obj match {
-      case JsonApiInfo(None, meta) if meta.isEmpty          => JsObject.empty
-      case JsonApiInfo(Some(version), meta) if meta.isEmpty => JsObject("version" -> version.toJson)
-      case JsonApiInfo(Some(version), meta)                 => JsObject("version" -> version.toJson, "meta" -> meta.toJson)
-    }
+  final implicit class JsonMapOps(val underlying: Map[String, JsValue]) extends AnyVal {
+    @inline def getAs[A](name: String)(implicit r: JsonReader[Option[A]]): Option[A] =
+      underlying.get(name).flatMap(_.convertTo[Option[A]])
 
-    override def read(json: JsValue): JsonApiInfo = {
-      val fields = json.asJsObject.fields
-      JsonApiInfo(version = fields.get("version").flatMap(_.convertTo[Option[String]]),
-                  meta = fields.get("meta").map(_.convertTo[MetaObject]).getOrElse(Map.empty))
-    }
+    @inline def getAs[A](name: Symbol)(implicit r: JsonReader[Option[A]]): Option[A] = getAs(name.name)
   }
 }
